+++ app/routes/application.js
@@ -6,25 +6,23 @@ import Analytics from 'ember-osf/mixins/analytics';
 
 export default Route.extend(Analytics, OSFAgnosticAuthRouteMixin, {
     i18n: inject(),
-    afterModel: function() {
+    afterModel() {
         const availableLocales = this.get('i18n.locales').toArray();
         let locale;
 
         // Works in Chrome and Firefox (editable in settings)
         if (navigator.languages && navigator.languages.length) {
-            for (let lang of navigator.languages) {
+            for (const lang of navigator.languages) {
                 if (availableLocales.includes(lang)) {
                     locale = lang;
                     break;
                 }
             }
-        }
-        // Backup for Safari (uses system settings)
-        else if (navigator.language && availableLocales.includes(navigator.language)) {
+        } else if (navigator.language && availableLocales.includes(navigator.language)) {
+            // Backup for Safari (uses system settings)
             locale = navigator.language;
         }
 
-        if (locale)
-            this.set('i18n.locale', locale);
-    }
+        if (locale) { this.set('i18n.locale', locale); }
+    },
 });