+++ tests/integration/components/preprint-navbar-branded-test.js
@@ -2,7 +2,7 @@ import { moduleForComponent, test } from 'ember-qunit';
 import hbs from 'htmlbars-inline-precompile';
 
 moduleForComponent('preprint-navbar-branded', 'Integration | Component | preprint navbar branded', {
-  integration: true,
+    integration: true,
 });
 
 test('it renders', function(assert) {