+++ tests/integration/components/search-facet-provider-test.js
@@ -4,36 +4,35 @@ import hbs from 'htmlbars-inline-precompile';
 
 moduleForComponent('search-facet-provider', 'Integration | Component | search facet provider', {
     integration: true,
-    beforeEach: function() {
-        let osfProvider = {
+    beforeEach() {
+        const osfProvider = {
             doc_count: 99,
-            key: 'OSF'
+            key: 'OSF',
         };
-        let clinicalTrialsProvider = {
+        const clinicalTrialsProvider = {
             doc_count: 100,
-            key: 'ClinicalTrials.gov'
+            key: 'ClinicalTrials.gov',
         };
-        let otherProviders =  A([
+        const otherProviders = A([
             osfProvider,
-            clinicalTrialsProvider
+            clinicalTrialsProvider,
         ]);
 
         this.set('otherProviders', otherProviders);
         this.set('facet', { key: 'sources', title: 'Providers', component: 'search-facet-provider' });
         this.set('key', 'sources');
-        let noop = () => {};
+        const noop = () => {};
         this.set('noop', noop);
         this.set('activeFilters', { providers: [], subjects: [] });
-        this.set('filterReplace',  {'Open Science Framework': 'OSF'});
-    }
+        this.set('filterReplace', { 'Open Science Framework': 'OSF' });
+    },
 });
 
 test('preprint providers and counts are listed', function(assert) {
     this.render(hbs`{{search-facet-provider
-        key=key
-        options=facet
         updateFilters=(action noop)
         activeFilters=activeFilters
+        options=facet
         filterReplace=filterReplace
         otherProviders=otherProviders
     }}`);
@@ -42,9 +41,9 @@ test('preprint providers and counts are listed', function(assert) {
 });
 
 test('filterReplace looks up key in mapping', function(assert) {
-   let osfProvider = {
+    const osfProvider = {
         doc_count: 99,
-        key: 'Open Science Framework'
+        key: 'Open Science Framework',
     };
     this.set('otherProviders', A([osfProvider]));
     this.render(hbs`{{search-facet-provider
@@ -54,6 +53,6 @@ test('filterReplace looks up key in mapping', function(assert) {
         activeFilters=activeFilters
         filterReplace=filterReplace
         otherProviders=otherProviders
-    }}`)
+    }}`);
     assert.equal(this.$('label')[0].innerText.trim(), 'OSF (99)');
 });