+++ tests/unit/routes/provider/index-test.js
@@ -1,15 +1,15 @@
 import { moduleFor, test } from 'ember-qunit';
 
 moduleFor('route:provider/index', 'Unit | Route | provider/index', {
-  // Specify the other units that are required for this test.
+    // Specify the other units that are required for this test.
     needs: [
         'route:index',
         'service:metrics',
-        'service:theme'
-    ]
+        'service:theme',
+    ],
 });
 
 test('it exists', function(assert) {
-  let route = this.subject();
-  assert.ok(route);
+    const route = this.subject();
+    assert.ok(route);
 });