+++ CHANGELOG.md
@@ -5,8 +5,15 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `ember-route-action-helper` for route closure actions
+- COS version of eslint
+
 ### Changed
 - Update `osf-style` to use the latest version with navbar changes
+
+### Fixed
+- Syntax errors from updating to COS version of eslint
 
 ## [0.6.12] - 2017-02-14
 ### Added